+++ branch-master/sqlmath_wasm.js
@@ -1,7 +1,7 @@
 /*jslint-disable*/
 // Copyright (c) 2021 Kai Zhu
 // SPDX-License-Identifier: MIT
-// 2025-12-29T01:06:00+0000
+// 2026-01-31T18:11:14+0000
 (function () {
 "use strict";