+++ branch-master/CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
 # Todo
-- betadog - Add function taskBacktest() to auto-remove features with low feature-importance.
 - sqlmath - Re-introduce sliding-discrete-fourier-transform to improve initial guess in sql-function WIN_SINEFIT2().
 - sqlite - Add similar error-handling from builtin-sql-function PERCENTILE() into custom-sql-function QUANTILE().
 - none
+
+# v2026.1.31
+- demo - Replace sector and subsector charts with subindustry chart.
+- chart - Make tooltip more transparent, so datapoints behind it are more visible.
+- chart - Cleanup chart _15_tradebot_tech_intra_1_month.
+- chart - Add intraday-technical-indicator tradebot_tech_intra.spy_zcl.
+- chart - Change time-scale of 1-month charts from 1-day to 15-minute/5-minute.
 
 # v2025.12.28
 - jslint - Upgrade to jslint-v2025.12.28.